requirements.txt
pandas>=2.2.3
datasets>=3.2.0
openai>=1.60.0
anthropic>=0.45.0
tqdm>=4.67.1
python-dotenv>=1.0.1
pytest>=7.0.0
seaborn>=0.13.2